Fix related to Using the JPEG-LS Encoding and Decoding Causes Occasional Application Crashes Fixes #191

+++ Native/Dicom.Imaging.Codec.JpegLS.cpp
@@ -39,7 +39,7 @@ EXPORT_Charls CharlsApiResultType JpegLSEncode(void* destination, size_t destina
     return JpegLsEncode(destination, destinationLength, bytesWritten, source, sourceLength, obj, errorMessage); 
 }
 
-EXPORT_Charls CharlsApiResultType JpegLSDecode(void * destination, int destinationLength, void* source, size_t sourceLength, JlsParameters* obj, char* errorMessage)
+EXPORT_Charls CharlsApiResultType JpegLSDecode(void * destination, size_t destinationLength, void* source, size_t sourceLength, JlsParameters* obj, char* errorMessage)
 {
     return JpegLsDecode(destination, destinationLength, source, sourceLength, obj, errorMessage);     
 }
